removed image_helper and updated roi numbering
